refactor(reflectutil): organization, namings, docs
- rename reflectutil.PathResolver -> reflectpath.Resolver
- rename methods accordingly
- split files by purpose: value, type, resolver
- add some docs
